# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 聆櫻聖境點歌系統 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /songs/{id}/lyrics:
    get:
      operationId: getSongLyrics
      tags: [lyrics]
      summary: Get lyrics for a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song lyrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongLyrics"
    patch:
      operationId: updateSongLyrics
      tags: [lyrics]
      summary: Save or update lyrics for a song (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongLyricsInput"
      responses:
        "200":
          description: Lyrics updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongLyrics"
    delete:
      operationId: deleteSongLyrics
      tags: [lyrics]
      summary: Remove lyrics from a song (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Lyrics removed
  /songs/{id}/lyrics/search:
    post:
      operationId: searchSongLyrics
      tags: [lyrics]
      summary: Search external lyrics providers for a song (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LyricsSearchInput"
      responses:
        "200":
          description: Lyrics search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LyricsSearchResult"
...
components:
  schemas:
...
    SongLyrics:
      type: object
      required: [songId]
      properties:
        songId:
          type: integer
        lyricsText:
          type: ["string", "null"]
        lyricsFormat:
          type: ["string", "null"]
        lyricsSource:
          type: ["string", "null"]
        lyricsUpdatedAt:
          type: ["string", "null"]
    SongLyricsInput:
      type: object
      required: [lyricsText, lyricsFormat, lyricsSource]
      properties:
        lyricsText:
          type: string
        lyricsFormat:
          type: string
          enum: [lrc, krc, yrc, plain]
        lyricsSource:
          type: string
          enum: [lrclib, netease, manual]
    LyricsSearchInput:
      type: object
      required: [provider]
      properties:
        provider:
          type: string
          enum: [lrclib, netease, auto]
        title:
          type: ["string", "null"]
        artist:
          type: ["string", "null"]
    LyricsSearchResultItem:
      type: object
      required: [provider, lyricsText, lyricsFormat, previewLines]
      properties:
        provider:
          type: string
        lyricsText:
          type: string
        lyricsFormat:
          type: string
        duration:
          type: ["integer", "null"]
        previewLines:
          type: array
          items:
            type: string
    LyricsSearchResult:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/LyricsSearchResultItem"
...