# lib/api-zod/src/generated/types/songLyrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 聆櫻聖境點歌系統 API
 * OpenAPI spec version: 0.1.0
 */

export interface SongLyrics {
  songId: number;
  /** @nullable */
  lyricsText?: string | null;
  /** @nullable */
  lyricsFormat?: string | null;
  /** @nullable */
  lyricsSource?: string | null;
  /** @nullable */
  lyricsUpdatedAt?: string | null;
}